Move MappableSelectItem to separate file
